+++ src/main/java/com/zeebra/domain/member/dto/MemberProfile.java
@@ -1,13 +1,11 @@
 package com.zeebra.domain.member.dto;
 
-import com.zeebra.domain.member.entity.Member;
-
 public record MemberProfile(
 	Long memberId,
 	String nickname,
 	String profileImage
 ) {
-	public static MemberProfile of(Member member) {
-		return new MemberProfile(member.getId(), member.getNickname(), member.getMemberImage());
+	public static MemberProfile of(MemberInfo member) {
+		return new MemberProfile(member.memberId(), member.nickname(), member.memberImage());
 	}
 }
