+++ src/main/java/com/zeebra/domain/notification/controller/NotificationController.java
@@ -1,7 +1,13 @@
 package com.zeebra.domain.notification.controller;
 
-import com.zeebra.domain.auth.dto.MemberInfo;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
 import com.zeebra.domain.auth.service.AuthService;
+import com.zeebra.domain.member.dto.MemberInfo;
+import com.zeebra.domain.member.service.MemberService;
 import com.zeebra.domain.notification.dto.NotificationResponse;
 import com.zeebra.domain.notification.dto.NotificationsResponse;
 import com.zeebra.domain.notification.service.NotificationService;
@@ -9,13 +15,10 @@ import com.zeebra.global.ApiResponse;
 import com.zeebra.global.ErrorCode.AuthErrorCode;
 import com.zeebra.global.exception.BusinessException;
 import com.zeebra.global.security.jwt.JwtProvider;
+
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("/api/notification")
@@ -23,11 +26,12 @@ import org.springframework.web.bind.annotation.RestController;
 public class NotificationController {
     private final NotificationService notificationService;
     private final AuthService authService;
+	private final MemberService memberService;
 
     @GetMapping("/all")
     public ApiResponse<NotificationsResponse> getNotifications(@AuthenticationPrincipal JwtProvider.JwtUserPrincipal principal) {
 
-        MemberInfo member = authService.findById(principal.getMemberId());
+        MemberInfo member = memberService.findById(principal.getMemberId());
         return ApiResponse.success(notificationService.getNotifications(member));
     }
